requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.6.0
streamlit>=1.12.0
pillow>=9.2.0
scikit-learn>=1.1.0
flask>=2.2.0
Pillow==9.5.0
matplotlib==3.7.1